% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_symbols}
\alias{binance_symbols}
\title{Symbols used in the accompanying paper}
\usage{
binance_symbols()
}
\value{
A character vector of Binance trading pairs.
}
\description{
Symbols used in the accompanying paper
}
\examples{
binance_symbols()
}
