% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.tpr_fit}
\alias{print.tpr_fit}
\title{Print a fitted model}
\usage{
\method{print}{tpr_fit}(x, ...)
}
\arguments{
\item{x}{An object of class `tpr_fit`.}

\item{...}{Ignored.}
}
\value{
`x`, invisibly.
}
\description{
Print a fitted model
}
\examples{
set.seed(1)
print(tpr_fit(matrix(rnorm(300), ncol = 3), rnorm(100)))
}
