% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.tpr_mc}
\alias{print.tpr_mc}
\title{Print a Monte Carlo study}
\usage{
\method{print}{tpr_mc}(x, ...)
}
\arguments{
\item{x}{An object of class `tpr_mc`.}

\item{...}{Ignored.}
}
\value{
`x`, invisibly.
}
\description{
Print a Monte Carlo study
}
\examples{
\donttest{
print(tpr_montecarlo(n = 150, sigma2 = 0.5, reps = 5, progress = FALSE))
}
}
